+++ src/app/app.component.ts
@@ -1,5 +1,4 @@
 import { Component } from '@angular/core';
-import { ROUTER_DIRECTIVES } from '@angular/router';
 
 import { ApiService } from './shared';
 
@@ -7,24 +6,17 @@ import { Wakanda } from './wakanda.service';
 
 import '../style/app.scss';
 
-/*
- * App Component
- * Top Level Component
- */
 @Component({
   selector: 'my-app', // <my-app></my-app>
-  providers: [ApiService, Wakanda],
-  directives: [...ROUTER_DIRECTIVES],
+  providers: [Wakanda],
   templateUrl: './app.component.html',
   styleUrls: ['./app.component.scss'],
 })
 export class AppComponent {
   url = 'https://github.com/preboot/angular2-webpack';
 
-  //constructor(private api: ApiService) {
-
-  wakandaClientVersion: string = undefined;
   constructor(private api: ApiService, public wakanda: Wakanda) {
     this.wakandaClientVersion = this.wakanda.wakandaClientVersion;
+    // Do something with api
   }
 }